# swagger-1.yml
openapi: 3.0.0
info:
  title: bookmeter
  description: bookmeterのパクリ
  version: "1.0"
servers:
  - url: http://localhost:5000
paths:
  /book/{ISBN}:
    get:
      summary: "書籍情報を探す"
      parameters:
        - in: path
          name: "ISBN"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "OK. Found."
          content:
            application/json:
              schema:
                type: "object"
              example:
                author: "Lubanovic, Bill"
                isbn: "9784873117386"
                permalink: "https://iss.ndl.go.jp/books/R100000074-I000686013-00"
                series: ""
                title: "入門Python3"
                volume: ""

        "404":
          description: "Not Found"
  /book/{ISBN}/update:
    get:
      summary: "書籍情報を強制更新"
      parameters:
        - in: path
          name: "isbn"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "OK"

  /status:
    get:
      summary: "読了状態を取得する"
      parameters:
        - in: query
          name: "uid"
          required: true
          schema:
            type: string

        - in: query
          name: "isbn"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Found"
          content:
            application/json:
              schema:
                type: "object"
              example:
                status: "read"
        "404":
          description: "Not Found"

    post:
      summary: "読了状態を記録する"
      parameters:
        - in: query
          name: "isbn"
          required: true
          schema:
            type: string
        - in: query
          name: "status"
          required: true
          schema:
            type: string
          example:
            - read
            - reading
            - unread
            - wish
        - in: query
          name: "uid"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "OK. Recorded."
        "403":
          description: "Invalid Token."
  /user/{uid}:
    get:
      summary: "読んだ本の一覧"
      parameters:
        - in: path
          name: "uid"
          required: true
          schema:
            type: string
        - in: query
          name: "status"
          schema:
            type: string
          example:
            - "read"
            - "reading"
            - "unread"
            - "wish"
      responses:
        "200":
          description: "Found"
          content:
            application/json:
              schema:
                type: "object"
                example:
                  - "isbn": "9784873117386"
                    "uid": "neso"
                    "status": "read"
                  - "isbn": "9784774183886"
                    "uid": "neso"
                    "status": "unread"
        "404":
          description: "Not Found"

  /search:
    get:
      summary: "書名から検索してISBNの一覧を表示する"
      parameters:
        - in: query
          required: true
          name: title
          schema:
            type: string
      responses:
        "200":
          description: "Search done"
          content:
            application/json:
              schema:
                type: "object"
                example:
                  - "author": "Lubanovic, Bill"
                    "isbn": "9784873117386"
                    "permalink": "https://iss.ndl.go.jp/books/R100000002-I026914699-00"
                    "publisher": "オライリー・ジャパン"
                    "series": ""
                    "title": "入門Python 3"
                    "volume": ""

                  - "author": "中久喜, 健司"
                    "isbn": "9784774183886"
                    "permalink": "https://iss.ndl.go.jp/books/R100000074-I000692947-00"
                    "publisher": "技術評論社"
                    "series": ""
                    "title": "科学技術計算のためのPython入門"
                    "volume": "1"

        "404":
          description: "search result none"